Move extension functions back to plugins
